feat: move calculation to a dedicated function
