+++ apps/namadillo/src/atoms/chain/atoms.ts
@@ -1,5 +1,4 @@
 import namada from "@namada/chains/chains/namada";
-import { singleUnitDurationFromInterval } from "@namada/utils/helpers";
 import { indexerApiAtom } from "atoms/api";
 import {
   defaultServerConfigAtom,
@@ -10,6 +9,7 @@ import { queryDependentFn } from "atoms/utils";
 import BigNumber from "bignumber.js";
 import { atomWithQuery } from "jotai-tanstack-query";
 import { ChainParameters, ChainSettings } from "types";
+import { calculateUnbondingPeriod } from "./functions";
 import { fetchChainParameters, fetchRpcUrlFromIndexer } from "./services";
 
 export const chainAtom = atomWithQuery<ChainSettings>((get) => {
@@ -73,30 +73,10 @@ export const chainParametersAtom = atomWithQuery<ChainParameters>((get) => {
     enabled: !!indexerUrl,
     queryFn: async () => {
       const parameters = await fetchChainParameters(api);
-
-      const unbondingPeriodInEpochs =
-        Number(parameters.unbondingLength) +
-        Number(parameters.pipelineLength) +
-        // + 1 because we unbonding period starts from the next epoch
-        1;
-      const minEpochDuration = Number(parameters.minDuration);
-      const minNumOfBlocks = Number(parameters.minNumOfBlocks);
-      const epochSwitchBlocksDelay = Number(parameters.epochSwitchBlocksDelay);
-
-      // Because epoch duration is in reality longer by epochSwitchBlocksDelay we have to account for that
-      const timePerBlock = minEpochDuration / minNumOfBlocks;
-      const realMinEpochDuration =
-        minEpochDuration + timePerBlock * epochSwitchBlocksDelay;
-
-      const unbondingPeriod = singleUnitDurationFromInterval(
-        0,
-        unbondingPeriodInEpochs * realMinEpochDuration
-      );
-
       return {
         ...parameters,
         apr: BigNumber(parameters.apr),
-        unbondingPeriod,
+        unbondingPeriod: calculateUnbondingPeriod(parameters),
       };
     },
   };
